#71 [REFACTOR] 폴더 경로 정리
