+++ src/main/java/com/example/trace/gpt/service/PostVerificationService.java
@@ -2,7 +2,7 @@ package com.example.trace.gpt.service;
 
 import com.example.trace.gpt.domain.Verification;
 import com.example.trace.gpt.dto.VerificationDto;
-import com.example.trace.post.dto.PostCreateDto;
+import com.example.trace.post.dto.post.PostCreateDto;
 
 public interface PostVerificationService {
     VerificationDto verifyPost(PostCreateDto postCreateDto,String providerId);
